Semantic or sentence splitting for normal chunking too
