+++ src/utils_langchain.py
@@ -18,8 +18,11 @@ from langchain.chains.summarize import map_reduce_prompt, LoadingCallable
 from langchain.chains.summarize.chain import _load_stuff_chain, _load_refine_chain, _load_map_reduce_chain
 from langchain.schema.language_model import BaseLanguageModel
 from langchain_community.embeddings import HuggingFaceHubEmbeddings
+from langchain_text_splitters import TextSplitter
 
-from src.utils import hash_file, get_sha, split_list, makedirs
+from src.enums import docs_joiner_default
+from src.utils import hash_file, get_sha, split_list, makedirs, flatten_list, get_token_count, get_docs_tokens, \
+    FakeTokenizer
 
 from langchain.callbacks.base import BaseCallbackHandler, Callbacks
 from langchain.schema import LLMResult
@@ -58,7 +61,7 @@ class StreamingGradioCallbackHandler(BaseCallbackHandler):
     def on_llm_new_token(self, token: str, **kwargs: Any) -> None:
         """Run on new LLM token. Only available when streaming is enabled."""
         if False and \
-        self.tgen0 is not None and self.max_time is not None and (time.time() - self.tgen0) > self.max_time:
+                self.tgen0 is not None and self.max_time is not None and (time.time() - self.tgen0) > self.max_time:
             if self.verbose:
                 print("Took too long in StreamingGradioCallbackHandler: %s" % (time.time() - self.tgen0), flush=True)
             self.text_queue.put(self.stop_signal)
@@ -100,7 +103,180 @@ class StreamingGradioCallbackHandler(BaseCallbackHandler):
             return value
 
 
-def _chunk_sources(sources, chunk=True, chunk_size=512, language=None, db_type=None):
+class H2OCharacterTextSplitter(RecursiveCharacterTextSplitter):
+    def __init__(
+            self,
+            separators: Optional[List[str]] = None,
+            keep_separator: bool = True,
+            is_separator_regex: bool = False,
+            **kwargs: Any,
+    ) -> None:
+        """Create a new TextSplitter."""
+        super().__init__(separators=separators, keep_separator=keep_separator, is_separator_regex=is_separator_regex,
+                         **kwargs)
+        self._separators = separators or ["\n\n", "\n", "  ", " ", ""]
+
+    @classmethod
+    def from_huggingface_tokenizer(cls, tokenizer: Any, **kwargs: Any) -> TextSplitter:
+        def _huggingface_tokenizer_length(text: str) -> int:
+            return get_token_count(text, tokenizer, add_special_tokens=False)
+
+        return cls(length_function=_huggingface_tokenizer_length, **kwargs)
+
+
+def select_docs_with_score(docs_with_score, top_k_docs, one_doc_size):
+    if top_k_docs > 0:
+        docs_with_score = docs_with_score[:top_k_docs]
+    elif one_doc_size is not None:
+        docs_with_score = [(docs_with_score[0][:one_doc_size], docs_with_score[0][1])]
+    else:
+        # do nothing
+        pass
+    return docs_with_score
+
+
+def split_merge_docs(docs_with_score, tokenizer=None, max_input_tokens=None, docs_token_handling=None,
+                     joiner=docs_joiner_default,
+                     non_doc_prompt='',
+                     do_split=True,
+                     hf_embedding_model=None,
+                     use_openai_embedding=False,
+                     verbose=False):
+    # group docs if desired/can to fill context to avoid multiple LLM calls or too large chunks
+    # only do first semantic split if have GPU
+    if hf_embedding_model and \
+            'model' in hf_embedding_model and \
+            not use_openai_embedding and \
+            hasattr(hf_embedding_model['model'], 'model_kwargs'):
+        do_first_semantic_split = hf_embedding_model['model'].model_kwargs.get('device') not in ['cpu']
+    else:
+        do_first_semantic_split = False
+
+    # NOTE: Could use joiner=\n\n, but if PDF and continues, might want just  full continue with joiner=''
+    # NOTE: assume max_input_tokens already processed if was -1 and accounts for model_max_len and is per-llm call
+    if max_input_tokens is not None:
+        max_input_tokens -= get_token_count(non_doc_prompt, tokenizer)
+
+    if docs_token_handling in ['chunk']:
+        return docs_with_score, 0
+    elif docs_token_handling in [None, 'split_or_merge']:
+        assert tokenizer
+        # see if need to split
+        # account for joiner tokens
+        joiner_tokens = get_token_count(joiner, tokenizer)
+        doc_chunk_size = max(64, min(max_input_tokens,
+                                     max(64, max_input_tokens - joiner_tokens * len(docs_with_score))))
+
+        if do_first_semantic_split and hf_embedding_model is not None and 'model' in hf_embedding_model:
+            # https://python.langchain.com/v0.1/docs/modules/data_connection/document_transformers/semantic-chunker/
+            from langchain_experimental.text_splitter import SemanticChunker
+            text_splitter0 = SemanticChunker(hf_embedding_model['model'])
+        else:
+            text_splitter0 = None
+
+        # skip split if not necessary, since expensive for some reason
+        text_splitter1 = H2OCharacterTextSplitter.from_huggingface_tokenizer(
+            tokenizer, chunk_size=doc_chunk_size, chunk_overlap=0,
+            separators=[". "],
+        )
+        text_splitter2 = H2OCharacterTextSplitter.from_huggingface_tokenizer(
+            tokenizer, chunk_size=doc_chunk_size, chunk_overlap=0,
+        )
+        # https://python.langchain.com/v0.1/docs/modules/data_connection/document_transformers/recursive_text_splitter/
+        text_splitter3 = H2OCharacterTextSplitter.from_huggingface_tokenizer(
+            tokenizer, chunk_size=doc_chunk_size, chunk_overlap=0,
+            separators=[
+                "\n\n",
+                "\n",
+                " ",
+                ".",
+                ",",
+                "\u200b",  # Zero-width space
+                "\uff0c",  # Fullwidth comma
+                "\u3001",  # Ideographic comma
+                "\uff0e",  # Fullwidth full stop
+                "\u3002",  # Ideographic full stop
+                "",
+            ],
+        )
+        text_splitters = dict(semantic=text_splitter0, sentence=text_splitter1, normal=text_splitter2,
+                              multilingual=text_splitter3)
+        text_splitters = {k: v for k, v in text_splitters.items() if v is not None}
+
+        did_split = False
+        for splitter_type, text_splitter in text_splitters.items():
+            tokens_before_split = [get_token_count(x + joiner, tokenizer) for x in
+                                   [x[0].page_content for x in docs_with_score]]
+
+            do_split &= any([x > max_input_tokens for x in tokens_before_split])
+            if not do_split:
+                break
+            did_split = True
+
+            if verbose:
+                print('tokens_before_split=%s' % tokens_before_split, flush=True)
+
+            [x[0].metadata.update(dict(docscore=x[1], doci=doci, ntokens=tokens_before_split[doci])) for doci, x in
+             enumerate(docs_with_score)]
+            docs = [x[0] for x in docs_with_score]
+            # only split those that need to be split, else recursive splitter goes too nuts and takes too long
+            docs_to_split = [x for x in docs if x.metadata['ntokens'] > doc_chunk_size]
+            docs_to_not_split = [x for x in docs if x.metadata['ntokens'] <= doc_chunk_size]
+            docs_split_new = flatten_list([text_splitter.split_documents([x]) for x in docs_to_split])
+            docs_new = docs_to_not_split + docs_split_new
+            doci_new = [x.metadata['doci'] for x in docs_new]
+            # order back by doci
+            docs_new = [x for _, x in sorted(zip(doci_new, docs_new), key=lambda pair: pair[0])]
+            docs_with_score = [(x, x.metadata['docscore']) for x in docs_new]
+
+            if verbose:
+                tokens_after_split = [get_token_count(x + joiner, tokenizer) for x in
+                                      [x[0].page_content for x in docs_with_score]]
+                print('tokens_after_split=%s' % tokens_after_split, flush=True)
+
+            if splitter_type == 'sentence' and len(docs_with_score) > 1:
+                # puts '. ' on next end of chunk, re-attach to end of previous chunk
+                docs_with_score = [
+                    (Document(x[0].page_content[2 if xi > 0 else 0:] + '.', metadata=x[0].metadata), x[1]) for xi, x in
+                    enumerate(docs_with_score)]
+
+        docs_with_score_new = []
+        k = 0
+        while k < len(docs_with_score):
+            # means use max_input_tokens to ensure model gets no more than max_input_tokens each map
+            top_k_docs, one_doc_size, num_doc_tokens = \
+                get_docs_tokens(tokenizer,
+                                text_context_list=[x[0].page_content for x in docs_with_score[k:]],
+                                max_input_tokens=max_input_tokens)
+            docs_with_score1 = select_docs_with_score(docs_with_score[k:], top_k_docs, one_doc_size)
+            new_score = docs_with_score1[0][1]
+            new_page_content = joiner.join([x[0].page_content for x in docs_with_score1])
+            new_metadata = docs_with_score1[0][0].metadata.copy()
+            new_metadata['source'] = joiner.join(set([x[0].metadata['source'] for x in docs_with_score1]))
+            doc1 = Document(page_content=new_page_content, metadata=new_metadata)
+            docs_with_score_new.append((doc1, new_score))
+
+            if did_split:
+                assert one_doc_size is None, "Split failed: %s" % one_doc_size
+            elif one_doc_size is not None:
+                # chopped
+                assert top_k_docs == 1
+            assert top_k_docs >= 1
+            k += top_k_docs
+
+        tokens_after_merge = [get_token_count(x + joiner, tokenizer) for x in
+                              [x[0].page_content for x in docs_with_score_new]]
+        if verbose:
+            print('tokens_after_merge=%s' % tokens_after_merge, flush=True)
+
+        max_tokens_after_merge = max(tokens_after_merge) if tokens_after_merge else 0
+        return docs_with_score_new, max_tokens_after_merge
+    else:
+        raise ValueError("No such docs_token_handling=%s" % docs_token_handling)
+
+
+def _chunk_sources(sources, chunk=True, chunk_size=512, language=None, db_type=None,
+                   new_splitter=True, hf_embedding_model=None, use_openai_embedding=False, verbose=False):
     assert db_type is not None
 
     if not isinstance(sources, (list, tuple, types.GeneratorType)) and not callable(sources):
@@ -125,9 +301,30 @@ def _chunk_sources(sources, chunk=True, chunk_size=512, language=None, db_type=N
         else:
             separators = ["\n\n", "\n", " ", ""]
             keep_separator = False
-        splitter = RecursiveCharacterTextSplitter(chunk_size=chunk_size, chunk_overlap=0, keep_separator=keep_separator,
-                                                  separators=separators)
-        source_chunks = splitter.split_documents(sources)
+        if not new_splitter:
+            splitter = RecursiveCharacterTextSplitter(chunk_size=chunk_size, chunk_overlap=0,
+                                                      keep_separator=keep_separator,
+                                                      separators=separators)
+            source_chunks = splitter.split_documents(sources)
+        else:
+            try:
+                tokenizer = FakeTokenizer(model_max_length=max(20, chunk_size - 50), is_openai=True)
+                sources_with_score = [(x, 1) for x in sources]
+                source_chunks_with_score, max_tokens_after_merge = \
+                    split_merge_docs(sources_with_score, tokenizer=tokenizer,
+                                     max_input_tokens=chunk_size, non_doc_prompt='',
+                                     do_split=True,
+                                     hf_embedding_model=hf_embedding_model if not use_openai_embedding else None,
+                                     verbose=verbose)
+                source_chunks = [x[0] for x in source_chunks_with_score]
+            except BaseException:
+                if os.getenv('HARD_ASSERTS'):
+                    raise
+                print("Failed to split with new method, use old method: %s" % str(e))
+                splitter = RecursiveCharacterTextSplitter(chunk_size=chunk_size, chunk_overlap=0,
+                                                          keep_separator=keep_separator,
+                                                          separators=separators)
+                source_chunks = splitter.split_documents(sources)
 
         # currently in order, but when pull from db won't be, so mark order and document by hash
         [x.metadata.update(dict(chunk_id=chunk_id)) for chunk_id, x in enumerate(source_chunks)]
@@ -154,7 +351,7 @@ def _add_meta(docs1, file, headsize=50, filei=0, parser='NotSet', file_as_source
         file_extension = pathlib.Path(file).suffix
         hashid = hash_file(file)
     else:
-        file_extension = str(file)  # not file, just show full thing
+        file_extension = str(type(file))
         hashid = get_sha(file)
     doc_hash = str(uuid.uuid4())[:10]
     if not isinstance(docs1, (list, tuple, types.GeneratorType)):
@@ -183,6 +380,7 @@ def fix_json_meta(docs1):
 class H2OMapReduceDocumentsChain(MapReduceDocumentsChain):
     allow_map_1 = True
     which = 'map'
+
     def combine_docs(
             self,
             docs: List[Document],
