+++ src/version.py
@@ -1,1 +1,1 @@
-__version__ = "e65cd157863e45ef71516e62d064185d4ccade2a"
+__version__ = "9aa4ab14603c51481f4964bc311db2f7a4e19f23"
